requirements.txt
moviepy==1.0.3
python-dotenv
requests
flask
openai==0.28.0
openai-whisper
torch
#ffmpeg
#ImageMagick